refactor: split terminal host server commands
